pyproject.toml
[project]
name = "compass-bot"
version = "0.9.4"
description = "A Discord bot in Python"
authors = [
    { name = "Glass Ships", email = "contact@glass-ships.com" },
    { name = "fatcatnine", email = "fatcatnine@gmail.com" },
]
requires-python = ">=3.12, <4.0"
readme = "README.md"
license = "GPL-3.0-or-later"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    ### Core
    "discord-py",
    "loguru",
    "psutil",
    "rich",
    "ruff",
    "typer",
    ### Database
    "xata",
    ### For Commands
    "dateparser",
    "GitPython",
    "httpx",
    "pillow",
    "python-dateutil",
    "pytz",
    ### For Music
    "beautifulsoup4",
    "pynacl",
    "spotipy",
    "yt-dlp",
]

[project.optional-dependencies]
dev = ["mkdocs", "mkdocs-material", "mkdocstrings[python]", "pytest"]

[project.scripts]
compass = "compass.main:app"

[project.urls]
Homepage = "https://gitlab.com/glass-ships/compass-bot"
Documentation = "https://glass-ships.gitlab.io/compass-bot/"
Repository = "https://gitlab.com/glass-ships/compass-bot"
Issues = "https://gitlab.com/glass-ships/compass-bot/-/issues"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/compass"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = [
    "F541", # f-strings with no placeholders
]
# exclude = []


[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "--import-mode=importlib"
